+++ Microsoft.Toolkit.Uwp.UI.Controls/GridSplitter/GridSplitter.cs
@@ -18,6 +18,7 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
 
         private GridResizeDirection _resizeDirection;
         private GridResizeBehavior _resizeBehavior;
+        private GripperHoverWrapper _hoverWrapper;
 
         /// <summary>
         /// Gets the target parent grid from level
