+++ Microsoft.Toolkit.Uwp.UI.Controls/GridSplitter/GripperHoverWrapper.cs
@@ -1,0 +1,136 @@
+﻿using Windows.UI.Core;
+using Windows.UI.Xaml;
+using Windows.UI.Xaml.Input;
+
+namespace Microsoft.Toolkit.Uwp.UI.Controls
+{
+    internal class GripperHoverWrapper
+    {
+        private readonly GridSplitter.GridResizeDirection _gridSplitterDirection;
+
+        private CoreCursor _splitterPreviousPointer;
+        private CoreCursor _previousCursor;
+        private GridSplitter.GripperCursorType _gripperCursor;
+        private int _gripperCustomCursorResource;
+        private bool _isDragging;
+
+        internal GridSplitter.GripperCursorType GripperCursor
+        {
+            get
+            {
+                return _gripperCursor;
+            }
+
+            set
+            {
+                _gripperCursor = value;
+            }
+        }
+
+        internal int GripperCustomCursorResource
+        {
+            get
+            {
+                return _gripperCustomCursorResource;
+            }
+
+            set
+            {
+                _gripperCustomCursorResource = value;
+            }
+        }
+
+        public GripperHoverWrapper(UIElement element, GridSplitter.GridResizeDirection gridSplitterDirection, GridSplitter.GripperCursorType gripperCursor, int gripperCustomCursorResource)
+        {
+            _gridSplitterDirection = gridSplitterDirection;
+            _gripperCursor = gripperCursor;
+            _gripperCustomCursorResource = gripperCustomCursorResource;
+            element.PointerEntered += Element_PointerEntered;
+            element.PointerExited += Element_PointerExited;
+        }
+
+        private void Element_PointerExited(object sender, PointerRoutedEventArgs e)
+        {
+            if (_isDragging)
+            {
+                // if dragging don't update the curser just update the splitter cursor with the last window cursor,
+                // because the splitter is still using the arrow cursor and will revert to original case when drag completes
+                _splitterPreviousPointer = _previousCursor;
+            }
+            else
+            {
+                Window.Current.CoreWindow.PointerCursor = _previousCursor;
+            }
+        }
+
+        private void Element_PointerEntered(object sender, PointerRoutedEventArgs e)
+        {
+            // if not dragging
+            if (!_isDragging)
+            {
+                _previousCursor = _splitterPreviousPointer = Window.Current.CoreWindow.PointerCursor;
+                UpdateDisplayCursor();
+            }
+
+            // if dragging
+            else
+            {
+                _previousCursor = _splitterPreviousPointer;
+            }
+        }
+
+        private void UpdateDisplayCursor()
+        {
+            if (_gripperCursor == GridSplitter.GripperCursorType.Default)
+            {
+                if (_gridSplitterDirection == GridSplitter.GridResizeDirection.Columns)
+                {
+                    Window.Current.CoreWindow.PointerCursor = GridSplitter.ColumnsSplitterCursor;
+                }
+                else if (_gridSplitterDirection == GridSplitter.GridResizeDirection.Rows)
+                {
+                    Window.Current.CoreWindow.PointerCursor = GridSplitter.RowSplitterCursor;
+                }
+            }
+            else
+            {
+                var coreCursor = (CoreCursorType)((int)_gripperCursor);
+                if (_gripperCursor == GridSplitter.GripperCursorType.Custom)
+                {
+                    if (_gripperCustomCursorResource > GridSplitter.GripperCustomCursorDefaultResource)
+                    {
+                        Window.Current.CoreWindow.PointerCursor = new CoreCursor(coreCursor, (uint)_gripperCustomCursorResource);
+                    }
+                }
+                else
+                {
+                    Window.Current.CoreWindow.PointerCursor = new CoreCursor(coreCursor, 1);
+                }
+            }
+        }
+
+        internal void SplitterManipulationStarted(object sender, ManipulationStartedRoutedEventArgs e)
+        {
+            var splitter = sender as GridSplitter;
+            if (splitter == null)
+            {
+                return;
+            }
+
+            _splitterPreviousPointer = splitter.PreviousCursor;
+            _isDragging = true;
+        }
+
+        internal void SplitterManipulationCompleted(object sender, ManipulationCompletedRoutedEventArgs e)
+        {
+            var splitter = sender as GridSplitter;
+            if (splitter == null)
+            {
+                return;
+            }
+
+            Window.Current.CoreWindow.PointerCursor = splitter.PreviousCursor = _splitterPreviousPointer;
+            _isDragging = false;
+        }
+    }
+}
