+++ Microsoft.Toolkit.Uwp.UI.Controls/Microsoft.Toolkit.Uwp.UI.Controls.csproj
@@ -56,6 +56,7 @@
     <Compile Include="GridSplitter\GridSplitter.Helper.cs" />
     <Compile Include="GridSplitter\GridSplitter.Options.cs" />
     <Compile Include="GridSplitter\GridSplitterGripper.cs" />
+    <Compile Include="GridSplitter\GripperHoverWrapper.cs" />
     <Compile Include="HamburgerMenu\MenuItems\HamburgerMenuItem.cs" />
     <Compile Include="HamburgerMenu\MenuItems\HamburgerMenuGlyphItem.cs" />
     <Compile Include="HamburgerMenu\MenuItems\HamburgerMenuItemCollection.cs" />
